Add random_password.py
Add old script that I used years ago to generate pseudo-random strings for use as passwords. While not technically "secure", I figured it was better than some of the previous passwords I used to come up with off the top of my head (which was more often than not just a variation of something I was already using).